+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Features:
   - Text component - set "section" to root element (CMS-5125)
   - Filtered content types in Create Content dialogue based on site-hierarchy (CMS-5109)
   - Automatic injection of component data attributes for layout, part, image and text components in Live Edit (CMS-5110)
+  - Media Types - Change FormIcon to use ContentResource.updateThumbnail instead of BlobResource.upload (CMS-4677)
 
 Refactoring: